Move common utils to a parent module
