+++ crates/module-system/sov-test-utils/src/evm/mod.rs
@@ -1,1 +1,22 @@
+use std::path::PathBuf;
+
+use ethers::contract::BaseContract;
+use ethers::core::abi::Abi;
+
+pub mod fake_uni;
 pub mod simple_storage;
+
+fn test_data_path() -> PathBuf {
+    let mut path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+    path.push("src");
+    path.push("evm");
+    path.push("test-data");
+    path.push("artifacts");
+    path
+}
+
+fn make_contract_from_abi(path: PathBuf) -> BaseContract {
+    let abi_json = std::fs::read_to_string(path).unwrap();
+    let abi: Abi = serde_json::from_str(&abi_json).unwrap();
+    BaseContract::from(abi)
+}
